# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Course & Training Management System API (ИПР №1)",
    "version": "1.0.0",
    "description": "Безопасный REST API для системы управления курсами, личным кабинетом студента, кабинетом преподавателя и динамическим RBAC (Role-Based Access Control). Спецификация сгенерирована из декларативных аннотаций Swaggo (// Method Godoc, @Summary, @Description, @Tags, @Param, @Success, @Failure, @Router) в контроллерах."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1 Base Endpoint"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Введите JWT Access токен (Bearer <token>)"
      }
    },
    "schemas": {
...
      "CreateInstructorRequest": {
        "type": "object",
        "required": [
          "user_id"
        ],
        "properties": {
          "user_id": {
            "type": "string",
            "example": "usr-uuid-example-1"
          },
          "bio": {
            "type": "string",
            "example": "Staff Backend Engineer, архитектор распределенных систем."
          },
          "specialization": {
            "type": "string",
            "example": "Golang, Highload, PostgreSQL, Docker"
          },
          "rating": {
            "type": "number",
            "example": 4.95
          }
        },
        "example": {
          "user_id": "usr-uuid-example-1",
          "bio": "Staff Backend Engineer, архитектор распределенных систем.",
          "specialization": "Golang, Highload, PostgreSQL, Docker",
          "rating": 4.95
        }
      },
      "UpdateInstructorRequest": {
        "type": "object",
        "properties": {
          "bio": {
            "type": "string",
            "example": "Главный технический руководитель направления Cloud-Native."
          },
          "specialization": {
            "type": "string",
            "example": "Golang, Kubernetes, Distributed Systems"
          },
          "rating": {
            "type": "number",
            "example": 5
          }
        },
        "example": {
          "bio": "Главный технический руководитель направления Cloud-Native.",
          "specialization": "Golang, Kubernetes, Distributed Systems",
          "rating": 5
        }
      },
...
      "InstructorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "instr-1"
              },
              "bio": {
                "type": "string",
                "example": "Senior Go Developer"
              },
              "specialization": {
                "type": "string",
                "example": "Golang, Highload"
              },
              "rating": {
                "type": "number",
                "example": 4.95
              },
              "user": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "example": "usr-1"
                  },
                  "full_name": {
                    "type": "string",
                    "example": "Алексей Архитекторов"
                  },
                  "email": {
                    "type": "string",
                    "example": "alex@mentor.ru"
                  }
                }
              }
            }
          }
        },
        "example": {
          "success": true,
          "data": {
            "id": "instr-1",
            "bio": "Senior Go Developer, 10 лет опыта в распределенных системах.",
            "specialization": "Golang, Distributed Systems, Docker",
            "rating": 4.95,
            "user": {
              "id": "usr-1",
              "full_name": "Алексей Архитекторов",
              "email": "alex@mentor.ru"
            }
          }
        }
      },
      "InstructorListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InstructorResponse"
            }
          }
        },
        "example": {
          "success": true,
          "data": [
            {
              "id": "instr-1",
              "specialization": "Golang, Highload",
              "rating": 4.95,
              "user": {
                "full_name": "Алексей Архитекторов",
                "email": "alex@mentor.ru"
              }
            }
          ]
        }
      },
      "InstructorStudentsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "instructor": {
                "type": "object"
              },
              "total_students": {
                "type": "integer",
                "example": 8
              },
              "enrollments": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              }
            }
          }
        },
        "example": {
          "success": true,
          "data": {
            "total_students": 2,
            "enrollments": [
              {
                "enrollment_id": "enr-1",
                "student_name": "Иван Иванов",
                "student_email": "student@example.com",
                "course_title": "Golang Architecture",
                "status": "confirmed"
              }
            ]
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "VALIDATION_ERROR"
              },
              "message": {
                "type": "string",
                "example": "Неверные параметры запроса."
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              }
            }
          }
        },
        "example": {
          "success": false,
          "error": {
            "code": "VALIDATION_ERROR",
            "message": "Неверные параметры запроса."
          }
        }
      }
    }
  },
  "paths": {
...
    "/instructors": {
      "get": {
        "tags": [
          "instructors"
        ],
        "summary": "Список преподавателей",
        "description": "Публичный список преподавателей с рейтингом и курсами",
        "parameters": [],
        "responses": {
          "200": {
            "description": "models.InstructorListResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstructorListResponse"
                },
                "example": {
                  "success": true,
                  "data": [
                    {
                      "id": "instr-1",
                      "specialization": "Golang, Highload",
                      "rating": 4.95,
                      "user": {
                        "full_name": "Алексей Архитекторов",
                        "email": "alex@mentor.ru"
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "instructors"
        ],
        "summary": "Создать профиль преподавателя",
        "description": "Создает профиль преподавателя (разрешение instructors:manage)",
        "parameters": [],
        "responses": {
          "201": {
            "description": "models.InstructorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstructorResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "instr-1",
                    "bio": "Senior Go Developer, 10 лет опыта в распределенных системах.",
                    "specialization": "Golang, Distributed Systems, Docker",
                    "rating": 4.95,
                    "user": {
                      "id": "usr-1",
                      "full_name": "Алексей Архитекторов",
                      "email": "alex@mentor.ru"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "models.ErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "message": "Неверные параметры запроса."
                  }
                }
              }
            }
          },
          "409": {
            "description": "models.ErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "message": "Неверные параметры запроса."
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "description": "Данные профиля",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInstructorRequest"
              },
              "example": {
                "user_id": "usr-uuid-example-1",
                "bio": "Staff Backend Engineer, архитектор распределенных систем.",
                "specialization": "Golang, Highload, PostgreSQL, Docker",
                "rating": 4.95
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/instructors/{id}": {
      "get": {
        "tags": [
          "instructors"
        ],
        "summary": "Профиль преподавателя по ID",
        "description": "Возвращает информацию о преподавателе",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Instructor ID",
            "schema": {
              "type": "string",
              "example": "c1a2b3c4-d5e6-7f8a-9b0c-1d2e3f4a5b6c"
            },
            "example": "c1a2b3c4-d5e6-7f8a-9b0c-1d2e3f4a5b6c"
          }
        ],
        "responses": {
          "200": {
            "description": "models.InstructorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstructorResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "instr-1",
                    "bio": "Senior Go Developer, 10 лет опыта в распределенных системах.",
                    "specialization": "Golang, Distributed Systems, Docker",
                    "rating": 4.95,
                    "user": {
                      "id": "usr-1",
                      "full_name": "Алексей Архитекторов",
                      "email": "alex@mentor.ru"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "models.ErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "message": "Неверные параметры запроса."
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "instructors"
        ],
        "summary": "Обновить профиль преподавателя",
        "description": "Обновление био, специализации и рейтинга (разрешение instructors:manage)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Instructor ID",
            "schema": {
              "type": "string",
              "example": "c1a2b3c4-d5e6-7f8a-9b0c-1d2e3f4a5b6c"
            },
            "example": "c1a2b3c4-d5e6-7f8a-9b0c-1d2e3f4a5b6c"
          }
        ],
        "responses": {
          "200": {
            "description": "models.InstructorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstructorResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "instr-1",
                    "bio": "Senior Go Developer, 10 лет опыта в распределенных системах.",
                    "specialization": "Golang, Distributed Systems, Docker",
                    "rating": 4.95,
                    "user": {
                      "id": "usr-1",
                      "full_name": "Алексей Архитекторов",
                      "email": "alex@mentor.ru"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "models.ErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "message": "Неверные параметры запроса."
                  }
                }
              }
            }
          },
          "404": {
            "description": "models.ErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "message": "Неверные параметры запроса."
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "description": "Данные профиля",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInstructorRequest"
              },
              "example": {
                "bio": "Главный технический руководитель направления Cloud-Native.",
                "specialization": "Golang, Kubernetes, Distributed Systems",
                "rating": 5
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/instructors/my/students": {
      "get": {
        "tags": [
          "instructors"
        ],
        "summary": "Список студентов преподавателя",
        "description": "Просмотр всех студентов, обучающихся на курсах преподавателя (разрешение instructors:view_students)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "models.InstructorStudentsResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstructorStudentsResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "total_students": 2,
                    "enrollments": [
                      {
                        "enrollment_id": "enr-1",
                        "student_name": "Иван Иванов",
                        "student_email": "student@example.com",
                        "course_title": "Golang Architecture",
                        "status": "confirmed"
                      }
                    ]
                  }
                }
              }
            }
          },
          "403": {
            "description": "models.ErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "message": "Неверные параметры запроса."
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}